add role based access

+++ src/controllers/tasks.controllers.js
@@ -5,7 +5,7 @@ import { apiResponse } from "../utils/apiResponse.utils.js";
 import { User } from "../models/users.models.js";
 
 const addTask=asyncHandler(async(req ,res,next)=>{
-    const {id,title,description,status,user}=req.body
+    const {id,title,description,status}=req.body
 
     if(!id){
         return next(apiError.badRequest(400,"id is required"))
@@ -16,16 +16,13 @@ const addTask=asyncHandler(async(req ,res,next)=>{
     else if(!status){
         return next(apiError.badRequest(400,"status is required"))
     }
-    else if(!user){
-        return next(apiError.badRequest(400,"user is required"))
-    }
 
     const task=await Task.create({
         id,
         title,
         description,
         status,
-        user
+        user:req.user._id
     })
 
     const createdtask=await Task.findById(task._id)
@@ -39,8 +36,23 @@ const addTask=asyncHandler(async(req ,res,next)=>{
 })
 
 const getTask=asyncHandler(async(req,res,next)=>{
-    const {id}=req.params
-    const task=await Task.findOne({id})
+    const task=await Task.aggregate([
+        {
+            $match:{user:req.user._id}
+        },
+        {$lookup:{
+            from:"users",
+            localField:"user",
+            foreignField:"_id",
+            as:"userDetails",
+            pipeline:[
+                {$project:{
+                    refreshToken:0,
+                    password:0
+                }}
+            ]
+        }}
+    ])
     
     if(!task){
         return next(apiError.notFound(404,"Task not found"))
@@ -51,7 +63,16 @@ const getTask=asyncHandler(async(req,res,next)=>{
 
 const updateTask=asyncHandler(async(req,res,next)=>{
     const {id}=req.params;
-    const {title,description,status,user}=req.body
+    const {title,description,status}=req.body
+
+    const task=await Task.findOne({id})
+    if(!task){
+        return next(apiError.notFound(404,"id not found!!"))
+    }
+
+    if(req.user.role !== "admin" && task.user.toString() !== req.user._id.toString()){
+        return next(apiError.unAuthorized(401,"you are not allowed to update this task"))
+    }
 
     const updatedtask=await Task.findOneAndUpdate(
         {id},
@@ -59,8 +80,7 @@ const updateTask=asyncHandler(async(req,res,next)=>{
             $set:{
                 title:title,
                 description:description,
-                status:status,
-                user:user
+                status:status
             }
         },
         {
@@ -80,6 +100,15 @@ const updateTask=asyncHandler(async(req,res,next)=>{
 const deleteTask=asyncHandler(async(req,res,next)=>{
     const {id}=req.params;
 
+    const task=await Task.findOne({id})
+    if(!task){
+        return next(apiError.notFound(404,"id not found!!"))
+    }
+
+    if(req.user.role !== "admin" && task.user.toString() !== req.user._id.toString()){
+        return next(apiError.unAuthorized(401,"you are not allowed to delete Task!!"))
+    }
+
     const deletedtask=await Task.findOneAndDelete({id})
 
     if(!deletedtask){
@@ -91,7 +120,15 @@ const deleteTask=asyncHandler(async(req,res,next)=>{
 })
 
 const getAllTasks=asyncHandler(async(req,res,next)=>{
-    const alltask=await Task.find()
+    //const alltask=await Task.find()
+    const alltask = await Task.aggregate([
+        {$lookup:{
+            from:"users",
+            localField:"user",
+            foreignField:"_id",
+            as:"userDetails",
+        }}
+    ])
 
     if(!alltask){
         return next(apiError.badRequest("Tasks are not fetched!!"))
